Fix bug in problem 64
It is somewhat suprising to me that the code produces the same output
with and without this change. I would expect the recursive call to
process_cf() with only three positional arguments to assign the empty
list to 'seen'. If anyone can shed some light on this, please let me
know.